requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.63
langchain-google-genai==2.0.3
langchain-openai==0.2.2
langchain-text-splitters==0.3.8
langgraph==0.2.28
langgraph-checkpoint==1.0.12
langsmith==0.1.147
marshmallow==3.26.1
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.109.1
orjson==3.11.4
packaging==24.2
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyparsing==3.2.5
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.44
sse-starlette==2.1.3
starlette==0.38.6
tavily-python==0.4.0
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.30.6
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
